fix: Resolve remaining compilation errors - down to 31 errors
- Fix proto v1 migration issues across storage engines
- Update metadata conversion patterns throughout codebase
- Fix type mismatches in SST, VIPER, cache systems
- Resolve SQL value enum usage and pattern matching
- Add missing imports and field completions
- Fix quantized_vector field types and HashMap conversions
- Update error handling and method resolution
- Major progress toward complete compilation success

+++ src/storage/engines/impls/sst/streaming_compaction.rs
@@ -291,7 +291,7 @@ impl StreamingCompactor {
                 // Check for tombstone (expired record)
                 let now = chrono::Utc::now().timestamp() as u32;
                 if let Some(expires_at) = merge_record.record.expires_at {
-                    if expires_at <= now {
+                    if expires_at <= now as i64 {
                         deleted_vector_ids.push(current_id.clone());
                         records_deduped += 1;
